Refactor class Integer comparison operator methods.

+++ src/network-integer.cpp
@@ -13,20 +13,17 @@ Network::Integer& Network::Integer::operator=(int t_value)
 
 bool Network::Integer::operator<(const Integer& t_integer) const
 {
-    return (static_cast<int>(m_value) <
-            static_cast<int>(t_integer.m_value));
+    return (m_value < static_cast<int>(t_integer.m_value));
 }
 
 bool Network::Integer::operator>(const Integer& t_integer) const
 {
-    return (static_cast<int>(m_value) >
-            static_cast<int>(t_integer.m_value));
+    return (m_value > static_cast<int>(t_integer.m_value));
 }
 
 bool Network::Integer::operator==(const Integer& t_integer) const
 {
-    return (static_cast<int>(m_value) ==
-            static_cast<int>(t_integer.m_value));
+    return (m_value == static_cast<int>(t_integer.m_value));
 }
 
 Network::Integer::operator int() const
